feat(rtl): rtl binder allow 64b io

+++ src/alkaid/codegen/rtl/common_source/binder_util.hh
@@ -9,7 +9,7 @@ constexpr bool _openmp = false;
 #endif
 
 template <typename CONFIG_T>
-std::enable_if_t<CONFIG_T::II != 0> _inference(int32_t *c_inp, int32_t *c_out, size_t n_samples) {
+std::enable_if_t<CONFIG_T::II != 0> _inference(int64_t *c_inp, int64_t *c_out, size_t n_samples) {
     auto dut = std::make_unique<typename CONFIG_T::dut_t>();
 
     size_t clk_req = n_samples * CONFIG_T::II + (CONFIG_T::latency - CONFIG_T::II) + 1;
@@ -39,7 +39,7 @@ std::enable_if_t<CONFIG_T::II != 0> _inference(int32_t *c_inp, int32_t *c_out, s
 }
 
 template <typename CONFIG_T>
-std::enable_if_t<CONFIG_T::II == 0> _inference(int32_t *c_inp, int32_t *c_out, size_t n_samples) {
+std::enable_if_t<CONFIG_T::II == 0> _inference(int64_t *c_inp, int64_t *c_out, size_t n_samples) {
     auto dut = std::make_unique<typename CONFIG_T::dut_t>();
 
     for (size_t i = 0; i < n_samples; ++i) {
@@ -52,7 +52,7 @@ std::enable_if_t<CONFIG_T::II == 0> _inference(int32_t *c_inp, int32_t *c_out, s
 }
 
 template <typename CONFIG_T>
-void batch_inference(int32_t *c_inp, int32_t *c_out, size_t n_samples, size_t n_threads) {
+void batch_inference(int64_t *c_inp, int64_t *c_out, size_t n_samples, size_t n_threads) {
     if (n_threads > 1 || n_threads == 0) {
 #ifdef _OPENMP
         size_t min_samples_per_thread;
